tested get full cover

+++ build/lib/ppsAnalysis/main.py
@@ -18,6 +18,7 @@ import seaborn as sns
 import ppsAnalysis.alignment
 import ppsAnalysis.cluster
 import ppsAnalysis.yeast_variant_analysis
+import ppsAnalysis.human_variant_analysis
 import ppsAnalysis.logthis
 import matplotlib.pyplot as plt
 from matplotlib_venn import venn2, venn2_circles, venn2_unweighted
@@ -76,33 +77,30 @@ def variants_main(arguments):
         if jobs_finished:
             main_logger.info("Alignment jobs all finished")
 
-    # after all alignment jobs finish, check VCF files and parse vcf files
-    parse_vcf_files(output, file_list, arguments, orfs, main_logger)
+    if arguments.mode == "human":
+        # after all alignment jobs finish, check VCF files and parse vcf files
+        parse_vcf_files_human(output, file_list, arguments, orfs, main_logger)
+    elif arguments.mode == "yeast":
+        # after all alignment jobs finish, check VCF files and parse vcf files
+        parse_vcf_files_yeast(output, file_list, orfs, main_logger)
 
 
-def parse_vcf_files(output, file_list, arguments, orfs, logger):
+def parse_vcf_files_human(output, file_list, arguments, orfs, logger):
     # for each sample, parse vcf files
     all_log = {"fastq_ID": [], "reads": [], "map_perc": []}
     genes_found = []
     all_genes_summary = pd.DataFrame([],
-                                     columns=["gene_ID", "gene_len", "total_rd", "avg_rd", "db", "count", "gene_name"])
+                                     columns=["gene_ID", "gene_len", "db", "count", "gene_name"])
     all_summary = os.path.join(output, "all_summary_subsetORF.csv")
     all_genes_summary.to_csv(all_summary, index=False)
     all_mut_df = []
     for f in file_list:
         if not f.endswith(".fastq.gz"): continue
-        if arguments.mode == "human":
-            # extract ID
-            fastq_ID = f.split("-")[-2]
-        elif arguments.mode == "yeast":
-            fastq_ID = f.split("_")[0]
-        else:
-            raise ValueError("Wrong mode")
+        fastq_ID = f.split(".")[0]
         sub_output = os.path.join(os.path.abspath(output), fastq_ID)
 
         # there should be only one log file in the dir
         log_file = glob.glob(f"{sub_output}/*.log")[0]
-
         if not os.path.isfile(log_file):
             logger.warning(f"log file does not exist: {log_file}")
             continue
@@ -115,27 +113,73 @@ def parse_vcf_files(output, file_list, arguments, orfs, logger):
                 if "alignment rate" in line:
                     perc_aligned = line.split("%")[0]
                     all_log["map_perc"].append(perc_aligned)
+        all_log["fastq_ID"] += [fastq_ID] * 2
+
+        # for each vcf file, get how many genes are fully aligned
+        # get only the subset that are in the group 
+        group_ID = fastq_ID.split("_")[-1][-1]
+        orfs_df = orfs[orfs["Pool group #"] == int(group_ID)]
+        print(orfs_df)
+        raw_vcf_file = os.path.join(sub_output, f"{fastq_ID}_group_spec_orfs_raw.vcf")
+        if os.path.isfile(raw_vcf_file):
+            # analysis of ORFs aligned to group specific reference
+            fully_covered, stats_list, mut_df = analysisHuman(raw_vcf_file, fastq_ID, orfs_df)
+            fully_covered_file = os.path.join(sub_output, "fully_covered_groupSpecORFs.csv")
+            fully_covered.to_csv(fully_covered_file, index=False)
+            print(fully_covered)
+            fully_covered.to_csv(all_summary, index=False, header=False, mode="a")
+            stats_list.append("groupSpecORFs")
+            genes_found.append(stats_list)
+            mut_df["sample"] = fastq_ID
+            all_mut_df.append(mut_df)
+        exit()
+
+def parse_vcf_files_yeast(output, file_list, orfs, logger):
+    # for each sample, parse vcf files
+    all_log = {"fastq_ID": [], "reads": [], "map_perc": []}
+    genes_found = []
+    all_genes_summary = pd.DataFrame([],
+                                     columns=["gene_ID", "gene_len", "db", "count", "gene_name"])
+    all_summary = os.path.join(output, "all_summary_subsetORF.csv")
+    all_genes_summary.to_csv(all_summary, index=False)
+    all_mut_df = []
+    for f in file_list:
+        if not f.endswith(".fastq.gz"): continue
+        fastq_ID = f.split("_")[0]
+        sub_output = os.path.join(os.path.abspath(output), fastq_ID)
+
+        # there should be only one log file in the dir
+        log_file = glob.glob(f"{sub_output}/*.log")[0]
+        if not os.path.isfile(log_file):
+            logger.warning(f"log file does not exist: {log_file}")
+            continue
+
+        # get information from the log file to make a summary log file for all the samples
+        with open(log_file, "r") as log_f:
+            for line in log_f:
+                if "reads;" in line:
+                    n_reads = line.split(" ")[0]
+                    all_log["reads"].append(n_reads)
+                if "alignment rate" in line:
+                    perc_aligned = line.split("%")[0]
+                    all_log["map_perc"].append(perc_aligned)
+
         all_log["fastq_ID"] += [fastq_ID] * 3
         # for each vcf file, get how many genes are fully aligned
-        if arguments.mode == "human":
-            # extract ID
-            pass
-        else:  # yeast
-            # first get the genes that are fully covered in the fastq files
-            orfs_df = orfs[orfs["plate"] == fastq_ID]
-            raw_vcf_file = os.path.join(sub_output, f"{fastq_ID}_L001_plateORFs_raw.vcf")
-            if os.path.isfile(raw_vcf_file):
-                # analysis of ORFs aligned to subgroup
-                fully_covered, stats_list, mut_df = analysisYeast(raw_vcf_file, fastq_ID, orfs_df)
-                fully_covered_file = os.path.join(sub_output, "fully_covered_plateORFs.csv")
-                fully_covered.to_csv(fully_covered_file, index=False)
-                fully_covered.to_csv(all_summary, index=False, header=False, mode="a")
-                db = fully_covered["db"].unique()
-                stats_list.append("plateORFs")
-                genes_found.append(stats_list)
-                mut_df["plate"] = fastq_ID
-                mut_df["db"] = db[0]
-                all_mut_df.append(mut_df)
+        orfs_df = orfs[orfs["plate"] == fastq_ID]
+        raw_vcf_file = os.path.join(sub_output, f"{fastq_ID}_L001_plateORFs_raw.vcf")
+        if os.path.isfile(raw_vcf_file):
+            # analysis of ORFs aligned to subgroup
+            fully_covered, stats_list, mut_df = analysisYeast(raw_vcf_file, fastq_ID, orfs_df)
+            fully_covered_file = os.path.join(sub_output, "fully_covered_plateORFs.csv")
+            fully_covered.to_csv(fully_covered_file, index=False)
+            fully_covered.to_csv(all_summary, index=False, header=False, mode="a")
+            db = fully_covered["db"].unique()
+            stats_list.append("plateORFs")
+            genes_found.append(stats_list)
+            mut_df["plate"] = fastq_ID
+            mut_df["db"] = db[0]
+            all_mut_df.append(mut_df)
 
     # process all log
     all_log = pd.DataFrame(all_log)
@@ -157,6 +201,7 @@ def parse_vcf_files(output, file_list, arguments, orfs, logger):
         "all_targeted_on_plate"]
     all_genes_stats.to_csv(genes_found_file, index=False)
 
+
 def read_yeast_csv(HIP_target_ORFs, other_target_ORFs):
     """
     Join HIP data and other data into one df, remove unwanted columns
@@ -177,6 +222,48 @@ def read_yeast_csv(HIP_target_ORFs, other_target_ORFs):
     return combined
 
 
+def read_human_csv(human91_ORFs):
+    humanallORF = pd.read_csv(human91_ORFs)
+    humanallORF = humanallORF[['orf_id', 'entrez_gene_id', 'Pool group #', 'entrez_gene_symbol', 'Mapped reads', 'Verified',
+                                '# mut']]
+    humanallORF = humanallORF.fillna(-1)
+    humanallORF["entrez_gene_id"] = humanallORF["entrez_gene_id"].astype(int)
+    humanallORF['orf_name'] = humanallORF['orf_id'].astype(str) + "_" + humanallORF['entrez_gene_id'].astype(str) + "_G0" + humanallORF['Pool group #'].astype(str) + "_" + humanallORF['entrez_gene_symbol'].astype(str)  
+
+    return humanallORF
+
+
+def analysisHuman(raw_vcf_file, fastq_ID, orfs_df):
+    """
+
+    """
+    analysis = ppsAnalysis.human_variant_analysis.humanAnalysis(raw_vcf_file, fastq_ID, orfs_df)
+    full_cover_genes, gene_dict, ref_dict = analysis.get_full_cover()
+
+    # all the genes with full coverage
+    n_fully_aligned = len(full_cover_genes.keys())
+    # all genes in ref fasta
+    n_ref = len(ref_dict.keys())
+    # all genes found in this fastq file
+    n_all_found = len(gene_dict.keys())
+
+    # save all the genes that are fully covered to the output folder
+    fully_covered = pd.DataFrame.from_dict(full_cover_genes, orient='index').reset_index()
+    fully_covered.columns = ["gene_ID", "gene_len"]
+
+    # merge with target orfs
+    merged_df = pd.merge(orfs_df, fully_covered, how="left", left_on="orf_name", right_on="gene_ID")
+
+    n_targeted = orfs_df.shape[0]
+    n_targeted_full = merged_df[~merged_df["gene_ID"].isnull()].shape[0]
+
+    # from fully aligned genes, select those with any mutations
+    stats_list = [fastq_ID, n_fully_aligned, n_all_found, n_targeted, n_targeted_full, n_ref]
+    print(stats_list)
+    return fully_covered, stats_list
+
+
+
 def analysisYeast(raw_vcf_file, fastq_ID, orfs_df):
     """
     Run yeast variants analysis, make files for each plate and save to corresponded dir
@@ -194,7 +281,7 @@ def analysisYeast(raw_vcf_file, fastq_ID, orfs_df):
 
     # save all the genes that are fully covered to the output folder
     fully_covered = pd.DataFrame.from_dict(full_cover_genes, orient='index').reset_index()
-    fully_covered.columns = ["gene_ID", "gene_len", "total_rd", "avg_rd"]
+    fully_covered.columns = ["gene_ID", "gene_len"]
     # split gene ID col 
     #fully_covered["gene_ID"] = fully_covered["gene_ID"].str.replace(to_replace="-[A-G]", "A")
      
@@ -271,9 +358,10 @@ def check_args(arguments):
         other_target_ORFs = "/home/rothlab/rli/02_dev/06_pps_pipeline/target_orfs/other_targeted_ORFs.csv"
         orfs = read_yeast_csv(HIP_target_ORFs, other_target_ORFs)
     elif arguments.mode == "human":
-        orfs =""
+        human_91ORFs = "/home/rothlab/rli/02_dev/06_pps_pipeline/fasta/human_91/20161117_ORFeome91_seqs.csv"
+        orfs = read_human_csv(human_91ORFs)
     else:
-        exit()
+        raise ValueError("Please provide valid mode")
 
     return orfs
